Addressed Tru64 compile warning.
git-svn-id: http://svn.code.sf.net/p/mpip/code/trunk@181 b8bbd010-f11e-0410-ac07-ecf92694ee1a

+++ report.c
@@ -444,7 +444,7 @@ mpiPi_profile_print (FILE * fp)
                          av[i]->id,
                          j,
                          fileLenMax+4,
-                         sizeof(void*)*2,
+                         (int)sizeof(void*)*2,
                          av[i]->pc[j], 
                          funcLenMax,
                          av[i]->functname[j],
